add swipe
