+++ package.json
@@ -183,6 +183,9 @@
     "react-router": "^5.3.3",
     "react-router-breadcrumbs-hoc": "^4.1.0",
     "react-router-dom": "^5.3.3",
+    "react-swipeable-views": "^0.14.0",
+    "react-swipeable-views-core": "^0.14.0",
+    "react-swipeable-views-utils": "^0.14.0",
     "react-table": "^7.8.0",
     "react-virtualized-auto-sizer": "^1.0.5",
     "react-window": "^1.8.7",
@@ -226,6 +229,9 @@
     "@types/react-infinite-scroller": "^1.2.3",
     "@types/react-redux": "^7.1.24",
     "@types/react-router-dom": "^5.3.2",
+    "@types/react-swipeable-views": "^0.13.5",
+    "@types/react-swipeable-views-core": "^0.13.3",
+    "@types/react-swipeable-views-utils": "^0.13.7",
     "@types/react-table": "^7.7.12",
     "@types/react-virtualized-auto-sizer": "^1.0.1",
     "@types/react-window": "^1.8.5",
