+++ src/components/Layout/Layout.tsx
@@ -6,13 +6,12 @@ import { Header } from './Header/Header'
 import { MobileNavBar } from './Header/NavBar/MobileNavBar'
 import { SideNav } from './Header/SideNav'
 
-const pbProp = { base: '72px', md: 0 }
 const widthProp = { base: 'full', md: 'calc(100% - 93px)', '2xl': 'calc(100% - 384px)' }
 const paddingBottomProp = { base: 'calc(0 + env(safe-area-inset-bottom))', md: 0 }
 
 export const Layout: React.FC<ContainerProps> = ({ children, ...rest }) => {
   return (
-    <Flex margin='0 auto' pb={pbProp}>
+    <Flex margin='0 auto'>
       <SideNav />
       <Container
         as='main'
