+++ src/pages/Dashboard/Dashboard.tsx
@@ -1,20 +1,26 @@
-import type { StackDirection } from '@chakra-ui/react'
-import { Stack } from '@chakra-ui/react'
-import { memo, useMemo } from 'react'
+import type { FlexProps, StackDirection, TabProps } from '@chakra-ui/react'
+import { Flex, Stack, Tab, TabIndicator, TabList, Tabs, useMediaQuery } from '@chakra-ui/react'
+import { memo, useMemo, useState } from 'react'
 import { useTranslate } from 'react-polyglot'
 import { Route, Switch, useRouteMatch } from 'react-router'
+import SwipeableViews from 'react-swipeable-views'
+import { mod } from 'react-swipeable-views-core'
+import { type SlideRenderProps, virtualize } from 'react-swipeable-views-utils'
 import { Main } from 'components/Layout/Main'
 import { SEO } from 'components/Layout/Seo'
 import { NftTable } from 'components/Nfts/NftTable'
 import { useFetchOpportunities } from 'components/StakingVaults/hooks/useFetchOpportunities'
 import { RawText } from 'components/Text'
 import { useFeatureFlag } from 'hooks/useFeatureFlag/useFeatureFlag'
+import { isMobile } from 'lib/globals'
 import { Accounts } from 'pages/Accounts/Accounts'
 import { TransactionHistory } from 'pages/TransactionHistory/TransactionHistory'
+import { breakpoints } from 'theme/theme'
 
-import { DashboardHeader } from './components/DashboardHeader'
+import { DashboardHeader } from './components/DashboardHeader/DashboardHeader'
 import { DashboardSidebar } from './DashboardSidebar'
 import { EarnDashboard } from './EarnDashboard'
+import { MobileActivity } from './MobileActivity'
 import { Portfolio } from './Portfolio'
 import { RewardsDashboard } from './RewardsDashboard'
 import { WalletDashboard } from './WalletDashboard'
@@ -22,19 +28,104 @@ import { WalletDashboard } from './WalletDashboard'
 const direction: StackDirection = { base: 'column', xl: 'row' }
 const maxWidth = { base: 'full', lg: 'full', xl: 'sm' }
 const pageProps = { pt: 0 }
+const customTabActive = { color: 'text.base' }
+const customTabLast = { marginRight: 0 }
+const CustomTab = (props: TabProps) => (
+  <Tab
+    fontWeight='semibold'
+    color='text.subtle'
+    _selected={customTabActive}
+    px={0}
+    py={4}
+    mr={6}
+    _last={customTabLast}
+    {...props}
+  />
+)
+
+const ScrollView = (props: FlexProps) => (
+  <Flex
+    flexDir='column'
+    width='100vw'
+    pt='calc(var(--mobile-header-offset) + 1rem)'
+    pb='var(--mobile-nav-offset)'
+    height='100vh'
+    overflowY='auto'
+    {...props}
+  />
+)
+
+const VirtualizedSwipableViews = virtualize(SwipeableViews)
 
 export const Dashboard = memo(() => {
   const translate = useTranslate()
+  const [slideIndex, setSlideIndex] = useState(0)
+  const [isLargerThanMd] = useMediaQuery(`(min-width: ${breakpoints['md']})`, { ssr: false })
   const isDefiDashboardEnabled = useFeatureFlag('DefiDashboard')
   const { path } = useRouteMatch()
   const isNftsEnabled = useFeatureFlag('Jaypegz')
+  const appIsMobile = isMobile || !isLargerThanMd
 
   const { isLoading: isOpportunitiesLoading } = useFetchOpportunities()
 
+  const handleSlideChange = (index: number) => {
+    setSlideIndex(index)
+  }
+
+  const mobileTabs = useMemo(() => {
+    return (
+      <Tabs mx={6} index={slideIndex} variant='unstyled' onChange={handleSlideChange}>
+        <TabList>
+          <CustomTab>Overview</CustomTab>
+          <CustomTab>NFTs</CustomTab>
+          <CustomTab>Activity</CustomTab>
+        </TabList>
+        <TabIndicator mt='-1.5px' height='2px' bg='blue.500' borderRadius='1px' />
+      </Tabs>
+    )
+  }, [slideIndex])
+
   const dashboardHeader = useMemo(
-    () => <DashboardHeader isOpportunitiesLoading={isOpportunitiesLoading} />,
-    [isOpportunitiesLoading],
+    () => (
+      <DashboardHeader isOpportunitiesLoading={isOpportunitiesLoading} tabComponent={mobileTabs} />
+    ),
+    [isOpportunitiesLoading, mobileTabs],
   )
+
+  const slideRenderer = (props: SlideRenderProps) => {
+    const { index, key } = props
+    let content
+    switch (mod(index, 3)) {
+      case 0:
+        content = <WalletDashboard />
+        break
+      case 1:
+        content = <NftTable />
+        break
+      case 2:
+        content = <MobileActivity />
+        break
+      default:
+        content = null
+        break
+    }
+    return <ScrollView key={key}>{content}</ScrollView>
+  }
+
+  if (appIsMobile) {
+    return (
+      <Main headerComponent={dashboardHeader} pageProps={pageProps} pt={0} pb={0}>
+        <VirtualizedSwipableViews
+          index={slideIndex}
+          slideRenderer={slideRenderer}
+          slideCount={3}
+          overscanSlideBefore={1}
+          overscanSlideAfter={1}
+          onChangeIndex={handleSlideChange}
+        />
+      </Main>
+    )
+  }
 
   if (isDefiDashboardEnabled)
     return (
